ref: move cmd funcs to "commands" dir
